ex06 done
This commit introduces the Harl class, which provides a complaint filtering mechanism.

The Harl class has methods for different levels of complaints (DEBUG, INFO, WARNING, ERROR) which, when called, prints out canned messages. The level to complain at is given as a command line argument when running the program.
If an invalid command line argument is given, prints out an error message.